refactor: #6 파일명 대소문자 변경 및 import 수정
